remove CRTP
